+++ test/cram/cat/migrate.t
@@ -95,9 +95,6 @@ Linting of the original file fails because we used workitems
 
   $ okra lint -e -C admin eng1.workitems.md
   [OK]: eng1.workitems.md
-  File "eng1.workitems.md", line 3:
-  Warning: Invalid objective:
-           "Property-Based Testing for Multicore (#1090)" is a work-item. You should use its parent objective "Property-Based Testing for Multicore (#558)" instead.
 
 We rewrite the file using okra cat